[FIX] allow workflow signals to be restarted
This commit fixes two parts, I'm not sure if
the second part is needed without the first.

Essentially, the first part allows concu-
rrency errors during workflow transitions to
restart the transaction, instead of being
thrown as an ValueError.

The second part fixes a problem when the
first (failed) transaction leaves references
to its database cursor behind, which then
blow up because it's already closed.